==> tdi_workshop/notebooks/02_ggplot-weatherdata.qmd <==
---
title: "How about that weather?"
subtitle: "GGPlot with Weather Data"
author: "Jeremy Mikecz"
format: html
editor: visual
---

# 1. Set up

## 1.1 Import necessary packages

```{r}
library(tidyverse)
```

## 1.2 import data

```{r}
year_avgs <- read.csv("../data/Hanover_temp-yearaverages.csv", row.names=1)

tail(year_avgs)
```

# 2. Graph Yearly Temperature Data

```{r}
ggplot(year_avgs |> filter(Year < 2025), aes(x=Year, y=T_yravg, color = type)) +
  geom_line() +
  geom_smooth()
```

Another way to do the same thing.

```{r}
ggplot() +
    geom_line(data=year_avgs |> filter (type=="maxtemp" & Year < 2024), aes(x=Year, y=T_yravg), color="red") +
    geom_line(data=year_avgs |> filter (type=="mintemp" & Year < 2024), aes(x=Year, y=T_yravg))
```

## 2.1 Adjust temperature data (long to wide), add color, and a rolling average

```{r}
year_avgs_wide <- year_avgs |>
    pivot_wider(names_from = type, values_from = T_yravg)
```

```{r}
library(zoo)
ggplot(year_avgs_wide |> filter(Year<2025), aes(x=Year)) +
    geom_line(aes(y=maxtemp), color="red", size=1, alpha=0.5) + 
    geom_line(aes(y=rollmean(maxtemp, 10, na.pad=TRUE)), linetype="dashed") +
    geom_line(aes(y=mintemp), color="blue", size=1, alpha=0.5)+
    geom_line(aes(y=rollmean(mintemp, 10, na.pad=TRUE)), linetype="dashed")
```

# 3. Daily Weather Data

## 3.1 Import data

```{r}
temps <- read.csv("../data/Hanover_temps2.csv")
```

## 3.2 Examine daily temperature data

```{r}
head(temps)
```

```{r}
colnames(temps)
```

## 3.3 Graph one year of daily temp data

```{r}
library(viridis)
ggplot(temps |> filter(Year == 2022), aes(x=daynum, y=max_diff20c_avg, fill=max_diff20c_avg)) +
    geom_bar(stat="identity") +
    scale_fill_viridis(option = "plasma")
    #scale_fill_gradient(low = "blue", high = "red")

```

## 3.4 Graph multiple years of daily temp data

```{r fig.width=14, fig.height = 60}
#options(repr.plot.width=30, repr.plot.height=50)
ggplot(temps |> filter(Year>2015), aes(x=daynum, y=max_diff20c_avg, fill=max_diff20c_avg)) +
    geom_bar(stat="identity") +
    facet_wrap(~Year, ncol=1) +
    scale_fill_viridis(option = "plasma")
```
